backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.4.6
google-generativeai==0.7.2
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9